pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "discord-plays"
version = "0.1.0"
description = "Chat-driven Steam controller emulator — Discord Plays"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3.2",
    "tomli>=2.0.1; python_version < '3.11'",
    "python-uinput>=0.11.2; sys_platform == 'linux'",
]

[project.optional-dependencies]
windows = [
    "vgamepad>=0.1.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
discord-plays = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["controller", "adapters"]
sources = [""]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
